Update location of contract authorization
