+++ lib/actions/OnboardingActions.js
@@ -38,11 +38,3 @@ export function setVerificationState (verificationState) {
     verificationState
   }
 }
-
-export function createCirclesPerson (address, signedTx) {
-  return {
-    type: CREATE_CIRCLES_PERSON,
-    address,
-    signedTx
-  }
-}
