+++ lib/components/Onboarding/OnboardingVerifyPhone.js
@@ -11,13 +11,10 @@ import {
   TouchableWithoutFeedback
 } from 'react-native'
 import { internationalFormat } from '../../utilities/phoneNumber'
-import { ethers } from 'ethers'
-import { SecureStore } from 'expo'
-import { customHttpProvider, HubContract } from 'circles-mobile/lib/utilities/blockchain'
 
 // Actions
 import { initSignUp, confirmSignUp } from '../../actions/AuthActions'
-import { setVerificationState, createCirclesPerson } from '../../actions/OnboardingActions'
+import { setVerificationState } from '../../actions/OnboardingActions'
 
 // import {
 //   verificationCall,
@@ -120,26 +117,6 @@ export class OnboardingVerifyPhone extends React.Component {
     if (verificationCode.length < 6) setVerificationState('incomplete')
     else if (verificationCode.length === 0) {
       setVerificationState('unverified')
-    }
-  }
-
-  async createCirclesPerson (name) {
-    let randomWallet = ethers.Wallet.createRandom()
-    let err = await SecureStore.setItemAsync('wallet', JSON.stringify(randomWallet))
-    if (!err) {
-      let connectedWallet = randomWallet.connect(customHttpProvider)
-      let transaction = {
-        nonce: await connectedWallet.getTransactionCount('pending'),
-        gasLimit: 1600000,
-        gasPrice: await customHttpProvider.getGasPrice(),
-        to: HubContract.address,
-        data: HubContract.interface.functions.signup.encode([name])
-      }
-      let tx = await connectedWallet.sign(transaction)
-      this.props.createCirclesPerson(randomWallet.address, tx)
-      this.props.navigation.push('Testnet')
-    } else {
-      // HANDLE THE ERROR
     }
   }
 
@@ -270,9 +247,6 @@ export const mapDispatchToProps = dispatch => {
     },
     setVerificationState: (verificationState) => {
       dispatch(setVerificationState(verificationState))
-    },
-    createCirclesPerson: (address, signedTx) => {
-      dispatch(createCirclesPerson(address, signedTx))
     }
     // verifyPhoneCode: (address, code) => {
     //   dispatch(verifyPhoneCode(address, code))
