+++ lib/sagas/AuthSaga.js
@@ -1,5 +1,9 @@
 // import { VERIFY_BY_SMS, VERIFY_PHONE_CODE, VERIFY_BY_PHONE, VERIFY_RECAPTCHA, VERIFY_FUNCAPTCHA } from 'circles-mobile/lib/constants/UportActionTypes'
 import { call, select, fork, put, take } from 'redux-saga/effects'
+import { Constants, SecureStore } from 'expo'
+import { ethers } from 'ethers'
+import { customHttpProvider, HubContract } from 'circles-mobile/lib/utilities/blockchain'
+import { push } from 'react-router-redux'
 import Auth from '@aws-amplify/auth'
 import API from '@aws-amplify/api'
 
@@ -11,8 +15,6 @@ import { hideLoadingSpinner, setAuthState, showLoadingSpinner } from 'circles-mo
 import { wipeOnboardingData, setVerificationState } from 'circles-mobile/lib/actions/OnboardingActions'
 import { SIGNIN_REQUEST, SIGNOUT_REQUEST, SIGNUP_REQUEST, SIGNUP_CONFIRM_REQUEST, SIGNIN_CONFIRM_REQUEST } from 'circles-mobile/lib/constants/AuthConstants'
 import { addCognitoUserData, addDbUserData } from 'circles-mobile/lib/actions/AuthActions'
-
-import { Constants } from 'expo'
 
 const logger = new ConsoleLogger('Auth Saga')
 
@@ -136,7 +138,8 @@ export function * confirmSignUp () {
     try {
       yield put(showLoadingSpinner())
       const response = yield Auth.confirmSignUp(phone, code)
-      if (response === 'SUCCESS') {        
+      if (response === 'SUCCESS') {
+
         yield put({ type: 'SIGNIN_REQUEST', data: action.data })
       } else throw Error('unexpected response:' + response)
     } catch (err) {
@@ -175,13 +178,33 @@ export function * initGetOrAddUser () {
     logger.debug('initGetOrAddUser GET response', response)
     if (!response.data) {
       response = yield API.post(apiName, path, myInit)
-      logger.debug('initGetOrAddUser POST response', response)      
+      logger.debug('initGetOrAddUser POST response', response)
     }
     yield put(addDbUserData(response.data))
   } catch (err) {
     logger.error(err)
   } finally {
     yield put(hideLoadingSpinner())
+  }
+}
+
+export function * createCirclesPerson (name) {
+  let randomWallet = ethers.Wallet.createRandom()
+  let err = yield SecureStore.setItemAsync('wallet', JSON.stringify(randomWallet))
+  if (!err) {
+    let connectedWallet = randomWallet.connect(customHttpProvider)
+    let transaction = {
+      nonce: yield connectedWallet.getTransactionCount('pending'),
+      gasLimit: 1600000,
+      gasPrice: yield customHttpProvider.getGasPrice(),
+      to: HubContract.address,
+      data: HubContract.interface.functions.signup.encode([name])
+    }
+    let tx = yield connectedWallet.sign(transaction)
+    this.props.createCirclesPerson(randomWallet.address, tx)
+    this.props.navigation.push('Testnet')
+  } else {
+    // HANDLE THE ERROR
   }
 }
