+++ lib/sagas/OnboardingSaga.js
@@ -20,11 +20,6 @@ import { hideLoadingSpinner, showLoadingSpinner } from 'circles-mobile/lib/actio
 
 const logger = new ConsoleLogger('Onboarding Saga')
 
-const NISABA_ROUTE = 'https://nisaba.uport.me'
-export const VERIFICATION_ROUTE = NISABA_ROUTE + '/api/v2/verify'
-export const CODE_VERIFICATION_ROUTE = VERIFICATION_ROUTE + '/check'
-export const REQUEST_PHONECALL_ROUTE = VERIFICATION_ROUTE + '/next'
-
 export function * saveProfileImage () {
   while (true) {
     let putResponse, getResponse, manipResult
